+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix(type): fix RequestHandler return type
 - refactor(errors): improve pathFilter error message
+- fix(logger-plugin): fix missing target port
 
 ## [v3.0.0](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.0)