+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - ci(package): npm package provenance
 - fix(logger-plugin): log target port when router option is used
 - refactor: fix circular dependencies
+- fix(fix-request-body): support '+json' content-type suffix
 
 ## [v3.0.0](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.0)