+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## next
+## [v3.0.2](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.2)
 
-- refactor(dependency): replace is-plain-obj with is-plain-object
-- chore(package): upgrade to eslint v9
+- refactor(dependency): replace is-plain-obj with is-plain-object ([#1031](https://github.com/chimurai/http-proxy-middleware/pull/1031))
+- chore(package): upgrade to eslint v9 ([#1032](https://github.com/chimurai/http-proxy-middleware/pull/1032))
+- fix(logger-plugin): handle undefined protocol and hostname ([#1036](https://github.com/chimurai/http-proxy-middleware/pull/1036))
 
 ## [v3.0.1](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.1)